refactor: schedules into schedules-rds.ts
